requirements.txt
setuptools>=65.0
wheel
pip>=23.0
packaging<25,>=16.8
absl-py==2.3.1
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
Automat==25.4.16
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.2
crashtest==0.4.1
croniter==2.0.5
cryptography==45.0.5
cssselect==1.3.0
curl_cffi==0.11.4
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.4.0
dnspython==2.7.0
dulwich==0.22.8
email_validator==2.2.0
executing==2.2.0
fake-useragent==1.5.1
Farama-Notifications==0.0.4
fastapi==0.111.0
fastapi-cli==0.0.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
flake8==7.1.0
Flask==2.3.3
flatbuffers==25.2.10
fonttools==4.58.3
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.73.1
gymnasium==0.29.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
installer==0.7.0
ipykernel==6.29.5
ipython
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.3.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.10.0
keyring==25.6.0
kiwisolver==1.4.8
Levenshtein==0.21.1
libclang==18.1.1
lightgbm==4.4.0
llvmlite==0.45.1
loguru==0.7.2
lxml==5.2.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.3.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
mss==10.0.0
multidict==6.6.3
multitasking==0.0.11
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx
newsapi-python==0.2.7
nltk==3.8.1
nodriver==0.47.0
notebook==7.4.4
notebook_shim==0.2.4
numba==0.62.1
numpy<=1.23.5
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.0
outcome==1.3.0.post0
overrides==7.7.0
pandas==2.0.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pbs-installer==2025.7.12
peewee==3.18.1
pillow==10.4.0
pkginfo==1.12.1.2
platformdirs==4.3.8
playwright==1.44.0
plotly==5.22.0
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
Protego==0.5.0
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
py==1.11.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
pyee==11.1.0
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.8.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-Levenshtein==0.21.1
python-multipart==0.0.20
python-whois==0.8.0
pytz==2024.1
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
qtconsole==5.6.1
QtPy==2.4.3
queuelib==1.8.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.14.8
rpds-py==0.25.1
safetensors==0.5.3
schedule==1.2.0
scikit-learn==1.3.0
scipy==1.15.3
Scrapy==2.13.3
seaborn==0.13.2
selenium==4.21.0
selenium-stealth==1.0.6
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
service-identity==24.2.0
shap==0.41.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.7
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.30
stable_baselines3==2.3.2
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
streamlit==1.35.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.18.1
textblob==0.19.0
textdistance==4.6.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.65.0
traitlets==5.14.3
transformers==4.53.2
trio==0.30.0
trio-websocket==0.12.2
trove-classifiers==2025.5.9.12
twilio==9.2.2
Twisted==25.5.0
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
undetected-chromedriver==3.5.4
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.30.1
validators==0.35.0
virtualenv==20.32.0
w3lib==2.3.1
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
win32_setctime==1.2.0
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.0.0
yarl==1.20.1
yfinance==0.2.63
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0